+++ src/client/java/tempeststudios/quickstacknearby/RecipeBookAwareButtonScreen.java
@@ -1,0 +1,6 @@
+package tempeststudios.quickstacknearby;
+
+public interface RecipeBookAwareButtonScreen {
+    default void quickstacknearby$updateButtonPositionsFromRecipeBookRender() {
+    }
+}